🐛 Fix: 모달 클릭 오류 수정 / 가이드 블로그 새창으로 이동
